Moved controls and scores to separate components
