+++ imports/ui/controls/controls.html
@@ -1,0 +1,20 @@
+<body>
+  {{> feed}}
+  {{> food}}
+  {{> reset}}
+</body>
+
+<template name="reset">
+  <button class="btn btn-success">Reset Game</button>
+</template>
+
+<template name="feed">
+  <p>Fireball has {{counter}} pieces of dragon food left.</p>
+  <p>{{mood}}</p>
+  <button class="btn btn-primary">Feed Fireball</button>
+</template>
+
+<template name="food">
+  <p><h2>Feed Fireball {{food}}</h2></p>
+  <button class="btn btn-primary">Switch Food</button><br/>
+</template>
